rename paasng.platform.cnative to paasng.platform.bkapp_model
