+++ apiserver/paasng/paas_wl/bk_app/cnative/specs/entities.py
@@ -74,7 +74,7 @@ class BkAppManifestProcessor:
                 self._patch_image(manifest, build.image)
             use_cnb = build.artifact_metadata.get("use_cnb", False)
 
-        # TODO: Replace below code with the manifest generated by `paasng.platform.cnative.bkapp_model`,
+        # TODO: Replace below code with the manifest generated by `paasng.platform.bkapp_model`,
         # Only inject the data that is not included in the manifest, such as ENVIRONMENT
         # related annotations and the built-in env variables.
